- rename pointScansUtils -> pointExtensionUtils
