+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/PointClouds.ts
@@ -21,7 +21,7 @@ import {
   pointCloudMaterial,
   pointCloudColorEncoding,
   POINT_CLOUD_REQUIRED_FIELDS,
-} from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/pointScanUtils";
+} from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/pointExtensionUtils";
 import type { RosObject, RosValue } from "@foxglove/studio-base/players/types";
 
 import { BaseUserData, Renderable } from "../Renderable";
