+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/pointExtensionUtils.ts
